requirements.txt
django==1.6.1
pyyaml==3.10
django-registration==1.0
